braille_converter_dpseek/requirements.txt
flask==3.0.0
python-docx==0.8.11
PyPDF2==3.0.1
textract==1.6.5
requests==2.31.0